Try components for sections
